gringotts/authentication/keysmith.py
from datetime import timedelta, datetime
from typing import Optional
from jose import JWTError, jwt

SECRET_KEY = "405980a6c0fe202c3ab87afc7da46ecf38cee7f0a811238a30986f43562a7183"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
